Adding hemiHists
